% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{Annotate,Seurat,CellAssignAnnotate-method}
\alias{Annotate,Seurat,CellAssignAnnotate-method}
\title{Annotate a Seurat object with CellAssignAnnotate (https://github.com/Irrationone/cellassign)}
\usage{
\S4method{Annotate}{Seurat,CellAssignAnnotate}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a CellAssignAnnotate object}

\item{...}{Additional arguments}
}
\value{
returns a data object with cell-wise annotation
}
\description{
Annotate a Seurat object with CellAssignAnnotate (https://github.com/Irrationone/cellassign)
}
\references{
Zhang, A.W., O’Flanagan, C., Chavez, E.A. et al. Probabilistic cell-type assignment of single-cell RNA-seq for tumor microenvironment profiling. Nat Methods 16, 1007–1015 (2019). https://doi.org/10.1038/s41592-019-0529-1
}
