% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{Annotate,Seurat,clustifyrAnnotate-method}
\alias{Annotate,Seurat,clustifyrAnnotate-method}
\title{Annotate a Seurat object with clustifyrAnnotate (https://github.com/rnabioco/clustifyr)}
\usage{
\S4method{Annotate}{Seurat,clustifyrAnnotate}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a clustifyrAnnotate object}

\item{...}{Additional arguments}
}
\value{
returns a data object with cell-wise annotation
}
\description{
Annotate a Seurat object with clustifyrAnnotate (https://github.com/rnabioco/clustifyr)
}
\references{
Fu R, Gillen AE, Sheridan RM, et al. clustifyr: an R package for automated single-cell RNA sequencing cluster classification. F1000Res. 2020;9:223. Published 2020 Apr 1. doi:10.12688/f1000research.22969.2
}
