% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate.r
\name{Annotate,Seurat,scCATCHAnnotate-method}
\alias{Annotate,Seurat,scCATCHAnnotate-method}
\title{Annotate a Seurat object with scCATCH (https://cran.r-project.org/web/packages/scCATCH/index.html)}
\usage{
\S4method{Annotate}{Seurat,scCATCHAnnotate}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a scCATCHAnnotate object}

\item{...}{Additional arguments}
}
\value{
returns a data object with cell-wise annotation
}
\description{
Annotate a Seurat object with scCATCH (https://cran.r-project.org/web/packages/scCATCH/index.html)
}
