% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{Annotate,Seurat,SCINAAnnotate-method}
\alias{Annotate,Seurat,SCINAAnnotate-method}
\title{Annotate a Seurat object with SCINA (https://github.com/jcao89757/SCINA)}
\usage{
\S4method{Annotate}{Seurat,SCINAAnnotate}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a SCINAAnnotate object}

\item{...}{Additional arguments}
}
\value{
returns a data object with cell-wise annotation
}
\description{
Annotate a Seurat object with SCINA (https://github.com/jcao89757/SCINA)
}
