% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{Annotate,Seurat,scMRMAAnnotate-method}
\alias{Annotate,Seurat,scMRMAAnnotate-method}
\title{Annotate a Seurat object with scMRMAAnnotate (https://github.com/JiaLiVUMC/scMRMA)}
\usage{
\S4method{Annotate}{Seurat,scMRMAAnnotate}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a CellAssignAnnotate object}

\item{...}{Additional arguments}
}
\value{
returns a data object with cell-wise annotation
}
\description{
Annotate a Seurat object with scMRMAAnnotate (https://github.com/JiaLiVUMC/scMRMA)
}
\references{
Jia Li, Quanhu Sheng, Yu Shyr, Qi Liu, scMRMA: single cell multiresolution marker-based annotation, Nucleic Acids Research, Volume 50, Issue 2, 25 January 2022, Page e7, https://doi.org/10.1093/nar/gkab931
}
