% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{Annotate,SeuratList,BaseAnnotate-method}
\alias{Annotate,SeuratList,BaseAnnotate-method}
\title{Annotate a SeuratList}
\usage{
\S4method{Annotate}{SeuratList,BaseAnnotate}(x, params, ...)
}
\arguments{
\item{x}{a SeuratList object}

\item{params}{a BaseAnnotate object}

\item{...}{Additional arguments}
}
\value{
returns a SeuratList object with cell-wise annotation
}
\description{
Annotate a SeuratList
}
