% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BasePreprocess-class}
\alias{BasePreprocess-class}
\title{The BasePreprocess class}
\description{
The BasePreprocess class
}
\section{Slots}{

\describe{
\item{\code{min_cells}}{the minimum number of cells that a gene should be expressed.}

\item{\code{min_genes}}{the minimum number of genes that a cell should express.}

\item{\code{norm_data}}{whether or not normalizing the data}

\item{\code{scaling}}{whether or not scaling the data}

\item{\code{norm_method}}{the normalization method}

\item{\code{scale_factor}}{the scaling factor}

\item{\code{numHVG}}{number of highly variable genes}

\item{\code{raw_assay}}{the raw assay field in a Seurat object}

\item{\code{batch}}{character name of batch in dataset metadata}
}}

