% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BBKNNMerge-class}
\alias{BBKNNMerge-class}
\title{The BBKNNMerge class}
\description{
The BBKNNMerge class
}
\section{Slots}{

\describe{
\item{\code{ridge_regress}}{whether or not performing ridge regression}

\item{\code{confounder_key}}{Confounder key}
}}

