% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPackage.R
\name{checkPackage}
\alias{checkPackage}
\title{Check package compatability and capability for running methods}
\usage{
checkPackage(package, version, language = "R")
}
\arguments{
\item{package}{string of package name to check}

\item{version}{version of package to request install if package is missing or below version number}

\item{language}{language that package is in}
}
\value{
returns NULL
}
\description{
Check package compatability and capability for running methods
}
