% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{Cluster,Seurat,KMeansCluster-method}
\alias{Cluster,Seurat,KMeansCluster-method}
\title{Cluster a Seurat object by K-means}
\usage{
\S4method{Cluster}{Seurat,KMeansCluster}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a KMeansCluster object}

\item{...}{Additional arguments}
}
\value{
returns a data object with clustering results in meta data
}
\description{
Cluster a Seurat object by K-means
}
