% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.r
\name{Cluster,Seurat,LeidenCluster-method}
\alias{Cluster,Seurat,LeidenCluster-method}
\title{Cluster a Seurat object by Leiden method}
\usage{
\S4method{Cluster}{Seurat,LeidenCluster}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a LeidenCluster object}

\item{...}{Additional arguments}
}
\value{
returns a data object with PCA embedding
}
\description{
Cluster a Seurat object by Leiden method
}
