% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.r
\name{Cluster,Seurat,LouvainCluster-method}
\alias{Cluster,Seurat,LouvainCluster-method}
\title{Cluster a Seurat object by Louvain method}
\usage{
\S4method{Cluster}{Seurat,LouvainCluster}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a LouvainCluster object}

\item{...}{Additional arguments}
}
\value{
returns a data object with clustering results in meta data
}
\description{
Cluster a Seurat object by Louvain method
}
