% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.r
\name{Cluster,Seurat,scCCESSSIMLRCluster-method}
\alias{Cluster,Seurat,scCCESSSIMLRCluster-method}
\title{Cluster a Seurat object by scCCESS SMILR (https://github.com/PYangLab/scCCESS)}
\usage{
\S4method{Cluster}{Seurat,scCCESSSIMLRCluster}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a LeidenCluster object}

\item{...}{Additional arguments}
}
\value{
returns a data object with clustering results in meta data
}
\description{
Cluster a Seurat object by scCCESS SMILR (https://github.com/PYangLab/scCCESS)
}
\references{
Geddes TA, Kim T, Nan L, Burchfield JG, Yang JYH, Tao D, et al. Autoencoder-based cluster ensembles for single-cell RNA-seq data analysis. BMC Bioinformatics. 2019;20:660.
}
