% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{Cluster,Seurat,SIMLRCluster-method}
\alias{Cluster,Seurat,SIMLRCluster-method}
\title{Cluster a Seurat object by SIMLR (https://www.bioconductor.org/packages/release/bioc/html/SIMLR.html)}
\usage{
\S4method{Cluster}{Seurat,SIMLRCluster}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a SIMLRCluster object}

\item{...}{Additional arguments}
}
\value{
returns a data object with clustering results in meta data
}
\description{
Cluster a Seurat object by SIMLR (https://www.bioconductor.org/packages/release/bioc/html/SIMLR.html)
}
\references{
Wang, B., Zhu, J., Pierson, E. et al. Visualization and analysis of single-cell RNA-seq data by kernel-based similarity learning. Nat Methods 14, 414–416 (2017). https://doi.org/10.1038/nmeth.4207

Wang B, Ramazzotti D, De Sano L, Zhu J, Pierson E, Batzoglou S. SIMLR: A Tool for Large-Scale Genomic Analyses by Multi-Kernel Learning. Proteomics. 2018 Jan;18(2). doi: 10.1002/pmic.201700232. PMID: 29265724.
}
