% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{Cluster,SeuratList,BaseCluster-method}
\alias{Cluster,SeuratList,BaseCluster-method}
\title{Cluster a SeuratList}
\usage{
\S4method{Cluster}{SeuratList,BaseCluster}(x, params, ...)
}
\arguments{
\item{x}{a SeuratList object}

\item{params}{a BaseCluster object}

\item{...}{Additional arguments}
}
\value{
returns a SeuratList object with clusters
}
\description{
Cluster a SeuratList
}
