% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet_detect.r
\name{DetectDoublet,Seurat,scDblFinderDoubletDetect-method}
\alias{DetectDoublet,Seurat,scDblFinderDoubletDetect-method}
\title{Doublet detection by scDblFinder (https://bioconductor.org/packages/release/bioc/html/scDblFinder.html)}
\usage{
\S4method{DetectDoublet}{Seurat,scDblFinderDoubletDetect}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a scDblFinderDoubletDetectobject}

\item{...}{Additional arguments}
}
\value{
returns a data object with doublet removed
}
\description{
Doublet detection by scDblFinder (https://bioconductor.org/packages/release/bioc/html/scDblFinder.html)
}
\references{
Germain P, Lun A, Macnair W, Robinson M (2021). “Doublet identification in single-cell sequencing data using scDblFinder.” f1000research. doi: 10.12688/f1000research.73600.1.
}
