% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet_detect.r
\name{DetectDoublet,Seurat,ScrubletDoubletDetect-method}
\alias{DetectDoublet,Seurat,ScrubletDoubletDetect-method}
\title{Doublet detection by Scrublet (https://github.com/swolock/scrublet)}
\usage{
\S4method{DetectDoublet}{Seurat,ScrubletDoubletDetect}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a ScrubletDoubletDetect object}

\item{...}{Additional arguments}
}
\value{
returns a data object with doublet removed
}
\description{
Doublet detection by Scrublet (https://github.com/swolock/scrublet)
}
\references{
Wolock SL, Lopez R, Klein AM. Scrublet: Computational Identification of Cell Doublets in Single-Cell Transcriptomic Data. Cell Syst. 2019 Apr 24;8(4):281-291.e9. doi: 10.1016/j.cels.2018.11.005. Epub 2019 Apr 3. PMID: 30954476; PMCID: PMC6625319.
}
