% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet_detect.r
\name{DetectDoublet,SeuratList,BaseDoubletDetect-method}
\alias{DetectDoublet,SeuratList,BaseDoubletDetect-method}
\title{Remove doublets from a SeuratList object}
\usage{
\S4method{DetectDoublet}{SeuratList,BaseDoubletDetect}(x, params, ...)
}
\arguments{
\item{x}{a SeuratList object}

\item{params}{a BaseDoubletDetect object}

\item{...}{Additional arguments}
}
\value{
returns a SeuratList object with doublets removed
}
\description{
Remove doublets from a SeuratList object
}
