% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_test.r
\name{DETest,Seurat,SeuratDETest-method}
\alias{DETest,Seurat,SeuratDETest-method}
\title{DE test of a Seurat objects by the default Seurat pipeline (https://satijalab.org/seurat/articles/de_vignette.html)}
\usage{
\S4method{DETest}{Seurat,SeuratDETest}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a SeuratDETest object}

\item{...}{Additional arguments}
}
\value{
a data frame with DE testing results
}
\description{
DE test of a Seurat objects by the default Seurat pipeline (https://satijalab.org/seurat/articles/de_vignette.html)
}
