% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_test.r
\name{DETest,SeuratList,BaseDETest-method}
\alias{DETest,SeuratList,BaseDETest-method}
\title{DE (differential expression) test for a SeuratList object}
\usage{
\S4method{DETest}{SeuratList,BaseDETest}(x, params, ...)
}
\arguments{
\item{x}{a SeuratList object}

\item{params}{a BaseDETest object}

\item{...}{Additional arguments}
}
\value{
returns a SeuratList object with DE results
}
\description{
DE (differential expression) test for a SeuratList object
}
