% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.r
\name{Embed,Seurat,PCAEmbed-method}
\alias{Embed,Seurat,PCAEmbed-method}
\title{Embed a Seurat object with PCA}
\usage{
\S4method{Embed}{Seurat,PCAEmbed}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a PCAEmbed object}

\item{...}{Additional arguments}
}
\value{
returns a data object with PCA embedding
}
\description{
Embed a Seurat object with PCA
}
