% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Embed.r
\name{Embed,Seurat,scCCESSSEmbed-method}
\alias{Embed,Seurat,scCCESSSEmbed-method}
\title{Embed a Seurat object with scCCESSS}
\usage{
\S4method{Embed}{Seurat,scCCESSSEmbed}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a scCCESSSEmbed object}

\item{...}{Additional arguments}
}
\value{
returns a data object with embedding results
}
\description{
Embed a Seurat object with scCCESSS
}
