% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.r
\name{Embed,Seurat,UMAPEmbed-method}
\alias{Embed,Seurat,UMAPEmbed-method}
\title{Embed a Seurat object with UMAP}
\usage{
\S4method{Embed}{Seurat,UMAPEmbed}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a UMAPEmbed object}

\item{...}{Additional arguments}
}
\value{
returns a data object with embedding results
}
\description{
Embed a Seurat object with UMAP
}
