% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Embed.r
\name{Embed,SeuratList,BaseEmbed-method}
\alias{Embed,SeuratList,BaseEmbed-method}
\title{Embed a SeuratList}
\usage{
\S4method{Embed}{SeuratList,BaseEmbed}(x, params, ...)
}
\arguments{
\item{x}{a SeuratList object}

\item{params}{a PCAEmbed object}

\item{...}{Additional arguments}
}
\value{
returns a SeuratList object with latent embedding
}
\description{
Embed a SeuratList
}
