% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.r
\name{ensemble,Seurat,EnsembleMerge-method}
\alias{ensemble,Seurat,EnsembleMerge-method}
\title{Merge Seurat objects by the EnsembleMerge method}
\usage{
\S4method{ensemble}{Seurat,EnsembleMerge}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a EnsembleMerge object}

\item{...}{Additional arguments}
}
\value{
returns a Seurat object of the integrated data
}
\description{
Merge Seurat objects by the EnsembleMerge method
}
