% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.r
\name{ensemble,Seurat-method}
\alias{ensemble,Seurat-method}
\title{Merge Seurat objects by the EnsembleMerge method}
\usage{
\S4method{ensemble}{Seurat}(x, methods = NULL, name = "Ensemble", ...)
}
\arguments{
\item{x}{a Seurat object}

\item{methods}{Name of constituent methods for ensembling}

\item{name}{Name of the ensembled results in the Seurat object}

\item{...}{Additional arguments}
}
\value{
returns a Seurat object of the integrated data
}
\description{
Merge Seurat objects by the EnsembleMerge method
}
