% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ensemble}
\alias{ensemble}
\title{The generic function of ensemble}
\usage{
ensemble(x, params, ...)
}
\arguments{
\item{x}{a data object}

\item{params}{a parameter object}

\item{...}{Other arguments}
}
\description{
The generic function of ensemble
}
