% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblemerge.R
\name{EnsembleMerge}
\alias{EnsembleMerge}
\title{Run enhsemblemerge functions}
\usage{
EnsembleMerge(
  data,
  methods = c("Seurat", "Harmony"),
  return = "SingleCellExperiment"
)
}
\arguments{
\item{x}{SummarizedExpirement object containing single cell counts matrix}
}
\value{
returns a SummarizedExperiment object of the integrated data
}
\description{
Run enhsemblemerge functions
}
