% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.r
\name{ensemblemerge_core}
\alias{ensemblemerge_core}
\title{The core ensemble function}
\usage{
ensemblemerge_core(params, data, a = 1, b = 1)
}
\arguments{
\item{params}{a EnsembleMergeParams object}

\item{data}{a Seurat object}

\item{a}{parameter for weight constituent method}

\item{b}{parameter for weight constituent method}
}
\value{
a list of two elememts: (1) the intergrated cell-cell neighboring graph, and (2) weight of each constituent method
}
\description{
The core ensemble function
}
