% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FastMNNMerge-class}
\alias{FastMNNMerge-class}
\title{The FastMNNMerge class}
\description{
The FastMNNMerge class
}
\section{Slots}{

\describe{
\item{\code{n_neighbors}}{number of neighbors used in calculating neighboring graph}
}}

