% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblemerge.R
\name{getMethods}
\alias{getMethods}
\title{Get valid integration methods}
\usage{
getMethods()
}
\value{
returns vector of valid names for integration methods
}
\description{
Get valid integration methods
}
