% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeighborGraph.R
\name{getNeighborGraph,BaseMerge,Seurat-method}
\alias{getNeighborGraph,BaseMerge,Seurat-method}
\title{Get a neighbor graph}
\usage{
\S4method{getNeighborGraph}{BaseMerge,Seurat}(params, data, ...)
}
\arguments{
\item{params}{a BaseMerge object}

\item{data}{a Seurat object}

\item{...}{Additional arguments}
}
\value{
returns a Seurat object with neighboring information
}
\description{
Get a neighbor graph
}
