% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeighborGraph.R
\name{getNeighborGraph,HarmonyParams-method}
\alias{getNeighborGraph,HarmonyParams-method}
\title{Get Neighbor graph}
\usage{
\S4method{getNeighborGraph}{HarmonyParams}(params, data)
}
\arguments{
\item{params}{HarmonyParams object generated by setParams()}

\item{data}{SingleCellExperiment object containing single cell counts matrix}
}
\value{
returns a SingleCellExperiment object of the integrated data
}
\description{
Get Neighbor graph
}
