% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeighborGraph.R
\name{getNeighborGraph,ScanoramaParams-method}
\alias{getNeighborGraph,ScanoramaParams-method}
\title{Get Neighbor graph}
\usage{
\S4method{getNeighborGraph}{ScanoramaParams}(
  params,
  data,
  normalization_method = "LogNormalize",
  latent = TRUE,
  ...
)
}
\arguments{
\item{params}{ScanoramaParams object generated by setParams()}

\item{data}{SingleCellExperiment object containing single cell counts matrix}
}
\value{
returns a SingleCellExperiment object of the integrated data
}
\description{
Get Neighbor graph
}
