% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeighborGraph.R
\name{getNeighborGraph,Seurat,EnsembleMerge-method}
\alias{getNeighborGraph,Seurat,EnsembleMerge-method}
\title{Get a neighbor graph}
\usage{
\S4method{getNeighborGraph}{Seurat,EnsembleMerge}(data, params, ...)
}
\arguments{
\item{data}{a Seurat object}

\item{params}{a EnsembleMerge object}

\item{...}{Additional arguments}
}
\value{
returns a Seurat object with neighboring information
}
\description{
Get a neighbor graph
}
