% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getNeighborGraph}
\alias{getNeighborGraph}
\title{The generic function of getNeighborGraph}
\usage{
getNeighborGraph(params, data, ...)
}
\arguments{
\item{params}{a parameter object}

\item{data}{a data object}

\item{...}{Other arguments}
}
\description{
The generic function of getNeighborGraph
}
