% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHarmony.R
\name{harmony_preprocess}
\alias{harmony_preprocess}
\title{Run Harmony functions}
\usage{
harmony_preprocess(
  x,
  normData = TRUE,
  Datascaling = TRUE,
  regressUMI = FALSE,
  norm_method = "LogNormalize",
  scale_factor = 10000,
  nfeatures = 300,
  npcs = 20,
  batch_label = "batchlb",
  celltype_label = "CellType"
)
}
\arguments{
\item{x}{SingleCellExperiment object containing single cell counts matrix}
}
\value{
returns a SingleCellExperiment object of the integrated data
}
\description{
Run Harmony functions
}
