% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.r
\docType{class}
\name{LeidenCluster-class}
\alias{LeidenCluster-class}
\title{The LeidenCluster class}
\description{
The LeidenCluster class
}
\section{Slots}{

\describe{
\item{\code{k_param}}{Defines k for the k-nearest neighbor algorithm}
}}

