% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLiger.R
\name{liger_preprocess}
\alias{liger_preprocess}
\title{Run LIGER functions}
\usage{
liger_preprocess(x, batch_label = "batchlb", var.thresh = 0.1)
}
\arguments{
\item{x}{SummarizedExpirement object containing single cell counts matrix}
}
\value{
returns a SummarizedExperiment object of the integrated data
}
\description{
Run LIGER functions
}
