% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{LigerHVG-class}
\alias{LigerHVG-class}
\title{LigerHVG}
\description{
parameters class for liger highly variable gene
}
\section{Slots}{

\describe{
\item{\code{var_threshold}}{theshold for variance of each gene in selecting variable features}
}}

