% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{LigerMerge-class}
\alias{LigerMerge-class}
\title{LigerMerge}
\description{
parameters class for liger merge
}
\section{Slots}{

\describe{
\item{\code{k}}{number of clustrs for liger merge}

\item{\code{nrep}}{number of restarts to perform}

\item{\code{lambda}}{Regularization parameter. Larger values penalize dataset-specific effects more strongly}
}}

