% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\docType{class}
\name{LouvainCluster-class}
\alias{LouvainCluster-class}
\title{The LouvainCluster class}
\description{
The LouvainCluster class
}
\section{Slots}{

\describe{
\item{\code{k_param}}{Defines k for the k-nearest neighbor algorithm}
}}

