% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{Merge,BaseMerge,Seurat-method}
\alias{Merge,BaseMerge,Seurat-method}
\title{Merge Seurat objects}
\usage{
\S4method{Merge}{BaseMerge,Seurat}(params, data, ...)
}
\arguments{
\item{params}{a BaseMerge object}

\item{data}{Seurat object containing single cell counts matrix}

\item{...}{Additional arguments}
}
\value{
returns a data object with integrated data
}
\description{
Merge Seurat objects
}
