% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{Merge,BaseMerge,SingleCellExperiment-method}
\alias{Merge,BaseMerge,SingleCellExperiment-method}
\title{Merge SingleCellExperiment objects}
\usage{
\S4method{Merge}{BaseMerge,SingleCellExperiment}(params, data, counts = "counts", ...)
}
\arguments{
\item{params}{a BaseMerge object}

\item{data}{a SingleCellExperiment}

\item{counts}{The assays field that contains the raw counts data}

\item{...}{Additional arguments}
}
\value{
returns a data object with integrated data
}
\description{
Merge SingleCellExperiment objects
}
