% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{Merge,BBKNNParams-method}
\alias{Merge,BBKNNParams-method}
\title{Merge SingleCellExperiment object}
\usage{
\S4method{Merge}{BBKNNParams}(params, data)
}
\arguments{
\item{params}{BBKNNParams object generated by setParams()}

\item{data}{SingleCellExperiment object containing single cell counts matrix}
}
\value{
returns a SingleCellExperiment object of the integrated data
}
\description{
Merge SingleCellExperiment object
}
