% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{Merge,EnsembleMergeParams,Seurat-method}
\alias{Merge,EnsembleMergeParams,Seurat-method}
\title{Merge Seurat objects by the EnsembleMerge method}
\usage{
\S4method{Merge}{EnsembleMergeParams,Seurat}(params, data, ...)
}
\arguments{
\item{params}{a EnsembleMergeParams object}

\item{data}{a Seurat object}

\item{...}{Additional arguments}
}
\value{
returns a data object (SingleCellExperiment or Seurat) of the integrated data
}
\description{
Merge Seurat objects by the EnsembleMerge method
}
