% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{Merge,HarmonyParams-method}
\alias{Merge,HarmonyParams-method}
\title{Merge SingleCellExperiment object}
\usage{
\S4method{Merge}{HarmonyParams}(params, data)
}
\arguments{
\item{params}{HarmonyParams object generated by setParams()}

\item{x}{SingleCellExperiment object containing single cell counts matrix}
}
\value{
returns a SingleCellExperiment object of the integrated data
}
\description{
Merge SingleCellExperiment object
}
