% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{Merge,LigerParams-method}
\alias{Merge,LigerParams-method}
\title{Merge SingleCellExperiment object}
\usage{
\S4method{Merge}{LigerParams}(params, data)
}
\arguments{
\item{params}{LigerParams object generated by setParams()}

\item{data}{SingleCellExperiment object containing single cell counts matrix}
}
\value{
returns a SingleCellExperiment object of the normalized/scaled data
}
\description{
Merge SingleCellExperiment object
}
