% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{Merge,Seurat,BaseMerge-method}
\alias{Merge,Seurat,BaseMerge-method}
\title{Merge Seurat objects}
\usage{
\S4method{Merge}{Seurat,BaseMerge}(data, params, ...)
}
\arguments{
\item{data}{Seurat object containing single cell counts matrix}

\item{params}{a BaseMerge object}

\item{...}{Additional arguments}
}
\value{
returns a data object with integrated data
}
\description{
Merge Seurat objects
}
