% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{Merge,Seurat,FastMNNMerge-method}
\alias{Merge,Seurat,FastMNNMerge-method}
\title{Merge Seurat objects}
\usage{
\S4method{Merge}{Seurat,FastMNNMerge}(data, params, ...)
}
\arguments{
\item{data}{Seurat object containing single cell counts matrix}

\item{params}{a FastMNNMerge object}

\item{...}{Additional arguments}
}
\value{
returns a Seurat object with integrated data
}
\description{
Merge Seurat objects
}
