% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{Merge,Seurat,SeuratMerge-method}
\alias{Merge,Seurat,SeuratMerge-method}
\title{Merge Seurat objects}
\usage{
\S4method{Merge}{Seurat,SeuratMerge}(data, params, ...)
}
\arguments{
\item{data}{a SeuratList object}

\item{params}{a SeuratMerge object}

\item{...}{Additional arguments}
}
\value{
returns a data object with integrated data
}
\description{
Merge Seurat objects
}
