% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.r
\name{Merge,SeuratList,BBKNNMerge-method}
\alias{Merge,SeuratList,BBKNNMerge-method}
\title{Merge Seurat objects by BBKNN (https://scanpy-tutorials.readthedocs.io/en/latest/integrating-data-using-ingest.html#BBKNN)}
\usage{
\S4method{Merge}{SeuratList,BBKNNMerge}(x, params, ...)
}
\arguments{
\item{x}{SeuratList object containing single cell counts matrix}

\item{params}{a BBKNNMerge object}

\item{...}{Additional arguments}
}
\value{
returns a Seurat object with integrated data
}
\description{
Merge Seurat objects by BBKNN (https://scanpy-tutorials.readthedocs.io/en/latest/integrating-data-using-ingest.html#BBKNN)
}
