% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.r
\name{Merge,SeuratList,ScanoramaMerge-method}
\alias{Merge,SeuratList,ScanoramaMerge-method}
\title{Merge Seurat objects}
\usage{
\S4method{Merge}{SeuratList,ScanoramaMerge}(x, params, ...)
}
\arguments{
\item{x}{SeuratList object containing single cell counts matrix}

\item{params}{a ScanoramaMerge object}

\item{...}{Additional arguments}
}
\value{
returns a Seurat object with integrated data
}
\description{
Merge Seurat objects
}
