% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.r
\name{Merge,SeuratList,scVIMerge-method}
\alias{Merge,SeuratList,scVIMerge-method}
\title{Merge Seurat objects by using scVI}
\usage{
\S4method{Merge}{SeuratList,scVIMerge}(x, params, ...)
}
\arguments{
\item{x}{SeuratList object containing single cell counts matrix}

\item{params}{a scVIMerge object}

\item{...}{Additional arguments}
}
\value{
returns a Seurat object with integrated data
}
\description{
Merge Seurat objects by using scVI
}
