% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{Merge,SummarizedExperiment,BaseMerge-method}
\alias{Merge,SummarizedExperiment,BaseMerge-method}
\title{Merge SummarizedExperiment objects}
\usage{
\S4method{Merge}{SummarizedExperiment,BaseMerge}(data, params, ...)
}
\arguments{
\item{data}{a SummarizedExperiment object}

\item{params}{a BaseMerge object}

\item{...}{Additional arguments}
}
\value{
returns a data object with integrated data
}
\description{
Merge SummarizedExperiment objects
}
