% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{Merge,SummarizedExperiment-method}
\alias{Merge,SummarizedExperiment-method}
\title{Merge SummarizedExperiment object}
\usage{
\S4method{Merge}{SummarizedExperiment}(
  x,
  split.by = NULL,
  batch_label = NULL,
  merge.which = NULL,
  method = "Seurat"
)
}
\arguments{
\item{x}{SummarizedExpirement object containing single cell counts matrix}
}
\value{
returns a SummarizedExperiment object of the integrated data
}
\description{
Merge SummarizedExperiment object
}
