% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{Merge,SummarizedExperiment,Params-method}
\alias{Merge,SummarizedExperiment,Params-method}
\title{Merge SummarizedExperiment objects}
\usage{
\S4method{Merge}{SummarizedExperiment,Params}(data, params, counts = "counts", ...)
}
\arguments{
\item{data}{a SummarizedExperiment object}

\item{params}{a MethodList object}

\item{counts}{the assay field for raw counts in a SingleCellExperiment object (default: 'counts')}

\item{...}{Additional arguments}
}
\value{
returns a data object with integrated data
}
\description{
Merge SummarizedExperiment objects
}
