% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{Merge}
\alias{Merge}
\title{Generic merge function}
\usage{
Merge(
  x,
  split.by = NULL,
  batch_label = NULL,
  cell_label = NULL,
  merge.which = NULL,
  method = "Seurat",
  ...
)
}
\description{
Generic merge function
}
