% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{Normalize,Seurat,scranNormalize-method}
\alias{Normalize,Seurat,scranNormalize-method}
\title{Normalize a Seurat objects by scran's devolution normalization (https://bioconductor.org/packages/devel/bioc/vignettes/scran/inst/doc/scran.html)}
\usage{
\S4method{Normalize}{Seurat,scranNormalize}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a scranNormalize object}

\item{...}{Additional arguments}
}
\value{
a Seurat object (if there is only one batch), or a SeuratList
}
\description{
Normalize a Seurat objects by scran's devolution normalization (https://bioconductor.org/packages/devel/bioc/vignettes/scran/inst/doc/scran.html)
}
\references{
L. Lun, A.T., Bach, K. & Marioni, J.C. Pooling across cells to normalize single-cell RNA sequencing data with many zero counts. Genome Biol 17, 75 (2016). https://doi.org/10.1186/s13059-016-0947-7
}
