% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{Normalize,Seurat,SCTransformNormalize-method}
\alias{Normalize,Seurat,SCTransformNormalize-method}
\title{Normalize a Seurat objects by SCTransform normalization}
\usage{
\S4method{Normalize}{Seurat,SCTransformNormalize}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a SCTransformNormalize object}

\item{...}{Additional arguments}
}
\value{
a Seurat object (if there is only one batch), or a SeuratList
}
\description{
Adopted from https://satijalab.org/seurat/articles/sctransform_vignette.html
}
\references{
Hafemeister, C., Satija, R. Normalization and variance stabilization of single-cell RNA-seq data using regularized negative binomial regression. Genome Biol 20, 296 (2019). https://doi.org/10.1186/s13059-019-1874-1
}
