% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{Normalize,Seurat,SeuratNormalize-method}
\alias{Normalize,Seurat,SeuratNormalize-method}
\title{Normalize a Seurat objects by the default Seurat pipeline}
\usage{
\S4method{Normalize}{Seurat,SeuratNormalize}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a SeuratNormalize object}

\item{...}{Additional arguments}
}
\value{
a Seurat object (if there is only one batch), or a SeuratList
}
\description{
Adopted from https://satijalab.org/seurat/articles/integration_introduction.html
}
