% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{Normalize,SeuratList,BaseNormalize-method}
\alias{Normalize,SeuratList,BaseNormalize-method}
\title{Normalize a SeuratList object}
\usage{
\S4method{Normalize}{SeuratList,BaseNormalize}(x, params, ...)
}
\arguments{
\item{x}{a SeuratList object}

\item{params}{a BaseNormalize object}

\item{...}{Additional arguments}
}
\value{
returns a SeuratList object with normalized read counts
}
\description{
Normalize a SeuratList object
}
