% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{Normalize}
\alias{Normalize}
\title{The generic function of Normalize}
\usage{
Normalize(x, params, ...)
}
\arguments{
\item{x}{a data object}

\item{params}{a parameter object}

\item{...}{Other arguments}
}
\description{
The generic function of Normalize
}
