% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc}
\alias{pbmc}
\title{293t_Jurkat cell lineage dataset with one batch with exclusively Jurkat cells}
\format{
A Seurat object frame with 500 cells and 500 features:
}
\source{
\url{https://s3.msi.umn.edu/skiex003/datasets/dataset=JC_benchmark_scRNAseq_version=2021219a/dataset_5.rds}
}
\usage{
pbmc
}
\description{
A dataset peripheral blood mononuclear cells in two distinct batches downsampled to 500 cells and 500 features
}
\keyword{datasets}
