% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.r
\name{Preprocess,ScanpyPreprocess,SingleCellExperiment-method}
\alias{Preprocess,ScanpyPreprocess,SingleCellExperiment-method}
\title{Preprocess a SingleCellExperiment object by the Scanpy pipeline}
\usage{
\S4method{Preprocess}{ScanpyPreprocess,SingleCellExperiment}(params, data, ...)
}
\arguments{
\item{params}{a ScanpyPreprocess object}

\item{data}{a SingleCellExperiment object}

\item{...}{Additional arguments}
}
\value{
returns a SingleCellExperiment object
}
\description{
Preprocess a SingleCellExperiment object by the Scanpy pipeline
}
