% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.r
\name{Preprocess,Seurat,BasePreprocess-method}
\alias{Preprocess,Seurat,BasePreprocess-method}
\title{Preprocess a Seurat objects}
\usage{
\S4method{Preprocess}{Seurat,BasePreprocess}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a BasePreprocess object}

\item{...}{Additional arguments}
}
\value{
a Seurat object (if there is only one batch), or a SeuratList
}
\description{
Preprocess a Seurat objects
}
