% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.r
\name{Preprocess,Seurat,ScanpyPreprocess-method}
\alias{Preprocess,Seurat,ScanpyPreprocess-method}
\title{Preprocess a Seurat object by the Scanpy pipeline}
\usage{
\S4method{Preprocess}{Seurat,ScanpyPreprocess}(data, params, ...)
}
\arguments{
\item{data}{a Seurat object}

\item{params}{a ScanpyPreprocess object}

\item{...}{Additional arguments}
}
\value{
returns a Seurat object
}
\description{
Preprocess a Seurat object by the Scanpy pipeline
}
