% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.r
\name{Preprocess,Seurat,SeuratPreprocess,BaseMerge-method}
\alias{Preprocess,Seurat,SeuratPreprocess,BaseMerge-method}
\title{Preprocess a Seurat objects by the Seurat pipeline}
\usage{
\S4method{Preprocess}{Seurat,SeuratPreprocess,BaseMerge}(data, params, merge, ...)
}
\arguments{
\item{data}{a Seurat object}

\item{params}{a SeuratPreprocess object}

\item{merge}{a BaseMerge object}

\item{...}{Additional arguments}
}
\value{
a Seurat object (if there is only one batch), or a list of Seurat objects
}
\description{
Adopted from https://satijalab.org/seurat/articles/integration_introduction.html
}
