% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.r
\name{Preprocess,Seurat,SeuratPreprocess-method}
\alias{Preprocess,Seurat,SeuratPreprocess-method}
\title{Preprocess a Seurat objects by the Seurat pipeline}
\usage{
\S4method{Preprocess}{Seurat,SeuratPreprocess}(data, params, ...)
}
\arguments{
\item{data}{a Seurat object}

\item{params}{a SeuratPreprocess object}

\item{...}{Additional arguments}
}
\value{
a Seurat object (if there is only one batch), or a SeuratList
}
\description{
Adopted from https://satijalab.org/seurat/articles/integration_introduction.html
}
