% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.r
\name{Preprocess,SeuratList,BasePreprocess-method}
\alias{Preprocess,SeuratList,BasePreprocess-method}
\title{Preprocess a SeuratList objects by the Seurat pipeline}
\usage{
\S4method{Preprocess}{SeuratList,BasePreprocess}(x, params, ...)
}
\arguments{
\item{x}{a SeuratList object}

\item{params}{a BasePreprocess object}

\item{...}{Additional arguments}
}
\value{
a SeuratList object
}
\description{
Preprocess a SeuratList objects by the Seurat pipeline
}
