% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.r
\name{Preprocess,SeuratPreprocess,Seurat-method}
\alias{Preprocess,SeuratPreprocess,Seurat-method}
\title{Preprocess a Seurat objects by the Seurat pipeline}
\usage{
\S4method{Preprocess}{SeuratPreprocess,Seurat}(params, data, ...)
}
\arguments{
\item{params}{a SeuratPreprocess object}

\item{data}{a Seurat object}

\item{...}{Additional arguments}
}
\value{
returns a Seurat object
}
\description{
Preprocess a Seurat objects by the Seurat pipeline
}
