% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.r
\name{Preprocess,SummarizedExperiment,SeuratPreprocess-method}
\alias{Preprocess,SummarizedExperiment,SeuratPreprocess-method}
\title{Preprocess a SummarizedExperiment objects by the Seurat pipeline}
\usage{
\S4method{Preprocess}{SummarizedExperiment,SeuratPreprocess}(data, params, counts = "counts", ...)
}
\arguments{
\item{data}{a Seurat object}

\item{params}{a SeuratPreprocess object}

\item{counts}{the assay field for raw counts in a SingleCellExperiment object (default: 'counts')}

\item{...}{Additional arguments}
}
\value{
a Seurat object (if there is only one batch), or a SeuratList
}
\description{
Adopted from https://satijalab.org/seurat/articles/integration_introduction.html
}
