% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMNN.R
\name{preprocess_MNN}
\alias{preprocess_MNN}
\title{Run Seurat V3 scaling and normalizing}
\usage{
preprocess_MNN(params, data)
}
\arguments{
\item{params}{UncorrectedParams object generated from setParams(method="Uncorrected", ...)}

\item{data}{SummarizedExpirement object containing single cell counts matrix}
}
\value{
returns a SummarizedExperiment object of the integrated data
}
\description{
Run Seurat V3 scaling and normalizing
}
