% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PythonPackage-class}
\alias{PythonPackage-class}
\title{PythonPackage}
\description{
The base params object for python packages
}
\section{Slots}{

\describe{
\item{\code{package_name}}{The package name}

\item{\code{package_version}}{The package version}
}}

