% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_map.r
\name{ReferenceMap,Seurat,Seurat,symphonyReferenceMap-method}
\alias{ReferenceMap,Seurat,Seurat,symphonyReferenceMap-method}
\title{Map a Seurat object (query) onto another Seruat object (atlas) by Symphony (https://github.com/immunogenomics/symphony)}
\usage{
\S4method{ReferenceMap}{Seurat,Seurat,symphonyReferenceMap}(query, atlas, params, ...)
}
\arguments{
\item{query}{a Seruat object}

\item{atlas}{a Seurat object}

\item{params}{a symphonyReferenceMap object}

\item{...}{Additional arguments}
}
\value{
returns a SeuratList object with query and atlas data with integrated dimension reduction
}
\description{
Map a Seurat object (query) onto another Seruat object (atlas) by Symphony (https://github.com/immunogenomics/symphony)
}
\references{
Kang, J.B., Nathan, A., Weinand, K. et al. Efficient and precise single-cell reference atlas mapping with Symphony. Nat Commun 12, 5890 (2021). https://doi.org/10.1038/s41467-021-25957-x
}
