% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_map.r
\name{ReferenceMap,Seurat,SeuratList,SeuratReferenceMap-method}
\alias{ReferenceMap,Seurat,SeuratList,SeuratReferenceMap-method}
\title{Map a Seurat object (query) onto another SeuratList object (atlas) by Seurat (https://satijalab.org/seurat/articles/integration_mapping.html)}
\usage{
\S4method{ReferenceMap}{Seurat,SeuratList,SeuratReferenceMap}(query, atlas, params, ...)
}
\arguments{
\item{query}{a Seurat object}

\item{atlas}{a SeuratList object  (with batches)}

\item{params}{a SeuratReferenceMap object}

\item{...}{Additional arguments}
}
\value{
returns a SeuratList object with query and atlas data with integrated dimension reduction
}
\description{
Map a Seurat object (query) onto another SeuratList object (atlas) by Seurat (https://satijalab.org/seurat/articles/integration_mapping.html)
}
