% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_map.r
\name{ReferenceMap,Seurat,SeuratList,symphonyReferenceMap-method}
\alias{ReferenceMap,Seurat,SeuratList,symphonyReferenceMap-method}
\title{Map a Seurat object (query) onto a SeruatList object (atlas) by Symphony (https://github.com/immunogenomics/symphony)}
\usage{
\S4method{ReferenceMap}{Seurat,SeuratList,symphonyReferenceMap}(query, atlas, params, ...)
}
\arguments{
\item{query}{a Seurat object}

\item{atlas}{a SeuratList object with multiple batches}

\item{params}{a symphonyReferenceMap object}

\item{...}{Additional arguments}
}
\value{
returns a SeuratList object with query and atlas data with integrated dimension reduction
}
\description{
Map a Seurat object (query) onto a SeruatList object (atlas) by Symphony (https://github.com/immunogenomics/symphony)
}
