% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_map.r
\name{ReferenceMap,SeuratList,SeuratList,scArchesReferenceMap-method}
\alias{ReferenceMap,SeuratList,SeuratList,scArchesReferenceMap-method}
\title{Map a SeuratList object (query) onto another SeuratList object (atlas) by scArches (https://scarches.readthedocs.io/en/latest/reference_building_from_scratch.html)}
\usage{
\S4method{ReferenceMap}{SeuratList,SeuratList,scArchesReferenceMap}(query, atlas, params, ...)
}
\arguments{
\item{query}{a SeuratList object with multiple batches}

\item{atlas}{a SeuratList object with multiple batches}

\item{params}{a scArchesReferenceMap object}

\item{...}{Additional arguments}
}
\value{
returns a SeuratList object with query and atlas data with integrated dimension reduction
}
\description{
Map a SeuratList object (query) onto another SeuratList object (atlas) by scArches (https://scarches.readthedocs.io/en/latest/reference_building_from_scratch.html)
}
\references{
Lotfollahi, M., Naghipourfar, M., Luecken, M.D. et al. Mapping single-cell data to reference atlases by transfer learning. Nat Biotechnol 40, 121–130 (2022). https://doi.org/10.1038/s41587-021-01001-7
}
