% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ReferenceMap}
\alias{ReferenceMap}
\title{The generic function of ReferenceMap}
\usage{
ReferenceMap(query, atlas, params, ...)
}
\arguments{
\item{query}{a data object}

\item{atlas}{a data object}

\item{params}{a parameter object}

\item{...}{Other arguments}
}
\description{
The generic function of ReferenceMap
}
