% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient_removal.r
\name{RemoveAmbientRNA,Seurat,decontXRemoveAmbientRNA-method}
\alias{RemoveAmbientRNA,Seurat,decontXRemoveAmbientRNA-method}
\title{Remove ambient RNAs by decontX (http://bioconductor.org/packages/release/bioc/vignettes/celda/inst/doc/decontX.html)}
\usage{
\S4method{RemoveAmbientRNA}{Seurat,decontXRemoveAmbientRNA}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a decontXRemoveAmbientRNA object}

\item{...}{Additional arguments}
}
\value{
returns a data object with ambient RNA removed
}
\description{
Remove ambient RNAs by decontX (http://bioconductor.org/packages/release/bioc/vignettes/celda/inst/doc/decontX.html)
}
\references{
Yang, S., Corbett, S.E., Koga, Y. et al. Decontamination of ambient RNA in single-cell RNA-seq with DecontX. Genome Biol 21, 57 (2020). https://doi.org/10.1186/s13059-020-1950-6
}
