% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient_removal.r
\name{RemoveAmbientRNA,Seurat,soupXRemoveAmbientRNA-method}
\alias{RemoveAmbientRNA,Seurat,soupXRemoveAmbientRNA-method}
\title{Remove ambient RNAs by soupX (https://github.com/constantAmateur/SoupX)}
\usage{
\S4method{RemoveAmbientRNA}{Seurat,soupXRemoveAmbientRNA}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a decontXRemoveAmbientRNA object}

\item{...}{Additional arguments}
}
\value{
returns a data object with ambient RNA removed
}
\description{
Remove ambient RNAs by soupX (https://github.com/constantAmateur/SoupX)
}
\references{
Matthew D Young, Sam Behjati, SoupX removes ambient RNA contamination from droplet-based single-cell RNA sequencing data, GigaScience, Volume 9, Issue 12, December 2020, giaa151, https://doi.org/10.1093/gigascience/giaa151
}
