% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient_removal.r
\name{RemoveAmbientRNA,SeuratList,BaseAmbientRNARemoval-method}
\alias{RemoveAmbientRNA,SeuratList,BaseAmbientRNARemoval-method}
\title{Remove ambient RNAs from a SeuratList object}
\usage{
\S4method{RemoveAmbientRNA}{SeuratList,BaseAmbientRNARemoval}(x, params, ...)
}
\arguments{
\item{x}{a SeuratList object}

\item{params}{a BaseAmbientRNARemoval object}

\item{...}{Additional arguments}
}
\value{
returns a SeuratList object with ambient RNA removed
}
\description{
Remove ambient RNAs from a SeuratList object
}
