% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RPackage-class}
\alias{RPackage-class}
\title{RPackage}
\description{
The base params object for R packages
}
\section{Slots}{

\describe{
\item{\code{package_name}}{The package name}

\item{\code{package_version}}{The package version}
}}

