% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMNN.R
\name{run_fastMNN}
\alias{run_fastMNN}
\title{Run fastMNN merging}
\usage{
run_fastMNN(params, data)
}
\arguments{
\item{params}{a FastMNNParams object}

\item{data}{a Seurat object}
}
\value{
returns a Seurat object with integrated data
}
\description{
Run fastMNN merging
}
