% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHarmony.R
\name{run_Harmony}
\alias{run_Harmony}
\title{Run Harmony functions}
\usage{
run_Harmony(params, data)
}
\arguments{
\item{params}{a HarmonyParams object}

\item{data}{a Seurat object}
}
\value{
returns a Seurat object with integrated data
}
\description{
Run Harmony functions
}
