% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLiger.R
\name{run_Liger}
\alias{run_Liger}
\title{Run Liger merging}
\usage{
run_Liger(params, data)
}
\arguments{
\item{params}{a LigerParams object}

\item{data}{a Seurat object}
}
\value{
returns a Seurat object of the integrated data
}
\description{
Run Liger merging
}
