% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runScanorama.R
\name{run_Scanorama}
\alias{run_Scanorama}
\title{Run Scanorama integration}
\usage{
run_Scanorama(params, data)
}
\arguments{
\item{params}{a ScanoramaParams object}

\item{data}{a Seurat object}
}
\value{
returns a Seurat object with integrated data
}
\description{
Run Scanorama integration
}
