% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCVI.R
\name{run_scVI}
\alias{run_scVI}
\title{Run scVI functions}
\usage{
run_scVI(params, data)
}
\arguments{
\item{params}{a scVIParams object}

\item{data}{a Seurat object}
}
\value{
returns a Seurat object with integrated data
}
\description{
Run scVI functions
}
