% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSeurat.R
\name{run_Seurat}
\alias{run_Seurat}
\title{Run Seurat functions}
\usage{
run_Seurat(params, data)
}
\arguments{
\item{params}{a SeuratParam object}

\item{data}{a list of Seurat objects or a SeuratList object}
}
\value{
returns a Seurat object with integrated data
}
\description{
Adopted from https://satijalab.org/seurat/articles/integration_introduction.html
}
