% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSMILE.R
\name{run_SMILE}
\alias{run_SMILE}
\title{Run Seurat V3 functions}
\usage{
run_SMILE(params, data)
}
\arguments{
\item{params}{SMILEParams object generated from setParams(method = "BBKNN") function}

\item{data}{SingleCellExperiment object containing single cell counts matrix}
}
\value{
returns a SummarizedExperiment object of the integrated data
}
\description{
Run Seurat V3 functions
}
