% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLiger.R
\name{RunOptimizeALS}
\alias{RunOptimizeALS}
\title{Run optimizeALS on a Seurat object}
\usage{
RunOptimizeALS(
  object,
  k,
  assay = NULL,
  split.by = "orig.ident",
  lambda = 5,
  thresh = 1e-06,
  max.iters = 30,
  reduction.name = "iNMF_raw",
  reduction.key = "riNMF_",
  nrep = 1,
  H.init = NULL,
  W.init = NULL,
  V.init = NULL,
  rand.seed = 1,
  print.obj = FALSE,
  ...
)
}
\arguments{
\item{object}{A merged Seurat object}

\item{k}{see rliger::optimizeALS}

\item{assay}{see rliger::optimizeALS}

\item{split.by}{Attribute for splitting, defaults to "orig.ident"}

\item{lambda}{see rliger::optimizeALS}

\item{thresh}{see rliger::optimizeALS}

\item{max.iters}{see rliger::optimizeALS}

\item{reduction.name}{Reduction name}

\item{reduction.key}{Reduction key}

\item{nrep}{see rliger::optimizeALS}

\item{H.init}{see rliger::optimizeALS}

\item{W.init}{see rliger::optimizeALS}

\item{V.init}{see rliger::optimizeALS}

\item{rand.seed}{Random seed}

\item{print.obj}{see rliger::optimizeALS}

\item{...}{Arguments passed to other methods}
}
\description{
Run optimizeALS on a Seurat object
}
