% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLiger.R
\name{RunQuantileNorm}
\alias{RunQuantileNorm}
\title{Run quantile_norm on a Seurat object}
\usage{
RunQuantileNorm(
  object,
  split.by = "orig.ident",
  reduction = "iNMF_raw",
  reduction.name = "iNMF",
  reduction.key = "iNMF_",
  quantiles = 50,
  ref_dataset = NULL,
  min_cells = 20,
  knn_k = 20,
  dims.use = NULL,
  do.center = FALSE,
  max_sample = 1000,
  eps = 0.9,
  refine.knn = TRUE,
  ...
)
}
\arguments{
\item{object}{A merged Seurat object}

\item{split.by}{Attribute for splitting, defaults to "orig.ident"}

\item{reduction}{Embedding name}

\item{reduction.name}{Reduction name}

\item{reduction.key}{Reduction key}

\item{quantiles}{see rliger::quantile_norm}

\item{ref_dataset}{see rliger::quantile_norm}

\item{min_cells}{see rliger::quantile_norm}

\item{knn_k}{see rliger::quantile_norm}

\item{dims.use}{see rliger::quantile_norm}

\item{do.center}{see rliger::quantile_norm}

\item{max_sample}{see rliger::quantile_norm}

\item{eps}{see rliger::quantile_norm}

\item{refine.knn}{see rliger::quantile_norm}

\item{...}{Arguments passed to other methods}
}
\description{
Run quantile_norm on a Seurat object
}
