% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{sc}
\alias{sc}
\title{on load initialization of reticulate environment}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
sc
}
\description{
on load initialization of reticulate environment
}
\keyword{datasets}
