% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{Score,SingleCellExperiment-method}
\alias{Score,SingleCellExperiment-method}
\title{score SingleCellExperiment object}
\usage{
\S4method{Score}{SingleCellExperiment}(x, batch_label = NULL, method = "kBET")
}
\arguments{
\item{x}{SingleCellExperiment object containing an integrated single cell counts matrix}
}
\value{
returns an output score statistic
}
\description{
score SingleCellExperiment object
}
