% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParams.R
\name{setParams}
\alias{setParams}
\title{Set parameters for integration}
\usage{
setParams(methods = "Seurat", ...)
}
\arguments{
\item{methods}{The method names (default: 'Seurat')}

\item{...}{Additional arguments}
}
\value{
If only one method is used, the output is a parameter object for each method; If more than one methods are used, the output is a `EnsembleMergeParams` object
}
\description{
Set parameters for integration
}
