% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSeurat.R
\name{seurat3_preprocess}
\alias{seurat3_preprocess}
\title{Run Seurat V3 functions}
\usage{
seurat3_preprocess(
  x,
  normData = TRUE,
  Datascaling = TRUE,
  min_cells = 10,
  min_genes = 300,
  norm_method = "LogNormalize",
  scale_factor = 10000,
  numVG = 300,
  nhvg = 2000,
  batch_label = "batchlb",
  celltype_label = "CellType",
  hvg = T,
  k.weight = 100
)
}
\arguments{
\item{x}{SummarizedExpirement object containing single cell counts matrix}
}
\value{
returns a SummarizedExperiment object of the integrated data
}
\description{
Run Seurat V3 functions
}
