% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SeuratHVG-class}
\alias{SeuratHVG-class}
\title{SeuratHVG}
\description{
Highly variable gene selection class for Seurat
}
\section{Slots}{

\describe{
\item{\code{numVG}}{number of variable genes for integration}

\item{\code{numHVG}}{number of highly variable genes to select for downstream analysis}
}}

