% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SeuratMerge-class}
\alias{SeuratMerge-class}
\title{SeuratMerge}
\description{
class for merging dataset with Seurat
}
\section{Slots}{

\describe{
\item{\code{npcs}}{number of principle components to use in dimension reduction}

\item{\code{seed}}{value to set for seed for umap}

\item{\code{dims}}{number of dimensions used in seurat preprocessing}

\item{\code{k.weight}}{weight for neighbor function}
}}

