% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SeuratParams-class}
\alias{SeuratParams-class}
\title{The SeuratParams class}
\description{
The SeuratParams class
}
\section{Slots}{

\describe{
\item{\code{altExp_names}}{name(s) for additional experiment object in dataset to keep in downstream analysis}
}}

