% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.r
\docType{class}
\name{SeuratPreprocess-class}
\alias{SeuratPreprocess-class}
\title{The SeuratPreprocess class}
\description{
The SeuratPreprocess class
}
\section{Slots}{

\describe{
\item{\code{selection.method}}{The gene selection method}

\item{\code{batchwise}}{whether or not performing batchwise data normalization and HVG selection}
}}

