% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\name{show,BaseMerge-method}
\alias{show,BaseMerge-method}
\title{Print class BaseMerge}
\usage{
\S4method{show}{BaseMerge}(object)
}
\arguments{
\item{object}{a SeuratPreprocess object}
}
\description{
Print class BaseMerge
}
