% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\name{show,BasePreprocess-method}
\alias{show,BasePreprocess-method}
\title{Print class BasePreprocess}
\usage{
\S4method{show}{BasePreprocess}(object)
}
\arguments{
\item{object}{a BasePreprocess object}
}
\description{
Print class BasePreprocess
}
