% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\name{show,BBKNNMerge-method}
\alias{show,BBKNNMerge-method}
\title{Print class BBKNNMerge}
\usage{
\S4method{show}{BBKNNMerge}(object)
}
\arguments{
\item{object}{a BBKNNMerge object}
}
\description{
Print class BBKNNMerge
}
