% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\name{show,EnsembleMerge-method}
\alias{show,EnsembleMerge-method}
\title{Print class EnsembleMerge}
\usage{
\S4method{show}{EnsembleMerge}(object)
}
\arguments{
\item{object}{a EnsembleMerge object}
}
\description{
Print class EnsembleMerge
}
