% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\name{show,FastMNNMerge-method}
\alias{show,FastMNNMerge-method}
\title{Print class FastMNNMerge}
\usage{
\S4method{show}{FastMNNMerge}(object)
}
\arguments{
\item{object}{a FastMNNMerge object}
}
\description{
Print class FastMNNMerge
}
