% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\name{show,HarmonyMerge-method}
\alias{show,HarmonyMerge-method}
\title{Print class HarmonyMerge}
\usage{
\S4method{show}{HarmonyMerge}(object)
}
\arguments{
\item{object}{a HarmonyMerge object}
}
\description{
Print class HarmonyMerge
}
