% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\name{show,LigerMerge-method}
\alias{show,LigerMerge-method}
\title{Print class LigerMerge}
\usage{
\S4method{show}{LigerMerge}(object)
}
\arguments{
\item{object}{a LigerMerge object}
}
\description{
Print class LigerMerge
}
