% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\name{show,SeuratMerge-method}
\alias{show,SeuratMerge-method}
\title{Print class SeuratMerge}
\usage{
\S4method{show}{SeuratMerge}(object)
}
\arguments{
\item{object}{a SeuratMerge object}
}
\description{
Print class SeuratMerge
}
