% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\name{show,SeuratPreprocess-method}
\alias{show,SeuratPreprocess-method}
\title{Print class SeuratPreprocess}
\usage{
\S4method{show}{SeuratPreprocess}(object)
}
\arguments{
\item{object}{a SeuratPreprocess object}
}
\description{
Print class SeuratPreprocess
}
