% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{UncorrectedParams-class}
\alias{UncorrectedParams-class}
\title{UncorrectedParams}
\description{
parameters class for uncorrected integration
}
\section{Slots}{

\describe{
\item{\code{vars_to_regress}}{list of names or name of metadata elements to regress}

\item{\code{hvg}}{boolean to check if hvg selection should be performed}
}}

