% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.r
\name{valid,Seurat,BaseDETest-method}
\alias{valid,Seurat,BaseDETest-method}
\title{Validate a Seurat object before DE test}
\usage{
\S4method{valid}{Seurat,BaseDETest}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a BaseDETest object}

\item{...}{Additional arguments}
}
\description{
Validate a Seurat object before DE test
}
