% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.r
\name{valid,Seurat,BaseMerge-method}
\alias{valid,Seurat,BaseMerge-method}
\title{Validate a Seurat object before merging}
\usage{
\S4method{valid}{Seurat,BaseMerge}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a BaseMerge object}

\item{...}{Additional arguments}
}
\description{
Validate a Seurat object before merging
}
