% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.r
\name{valid,Seurat,BaseNormalize-method}
\alias{valid,Seurat,BaseNormalize-method}
\title{Validate a Seurat object before normalization}
\usage{
\S4method{valid}{Seurat,BaseNormalize}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a BaseNormalize object}

\item{...}{Additional arguments}
}
\description{
Validate a Seurat object before normalization
}
