% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.r
\name{valid,Seurat,BasePreprocess-method}
\alias{valid,Seurat,BasePreprocess-method}
\title{Validate a Seurat object before preprocessing}
\usage{
\S4method{valid}{Seurat,BasePreprocess}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a BasePreprocess object}

\item{...}{Additional arguments}
}
\description{
Validate a Seurat object before preprocessing
}
