% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.r
\name{valid,Seurat,EnsembleMerge-method}
\alias{valid,Seurat,EnsembleMerge-method}
\title{Validate a Seurat object before ensembling}
\usage{
\S4method{valid}{Seurat,EnsembleMerge}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a EnsembleMerge object}

\item{...}{Additional arguments}
}
\description{
Validate a Seurat object before ensembling
}
