% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.r
\name{valid,SeuratList,BaseMerge-method}
\alias{valid,SeuratList,BaseMerge-method}
\title{Validate a SeuratList object before merging}
\usage{
\S4method{valid}{SeuratList,BaseMerge}(x, params, ...)
}
\arguments{
\item{x}{a SeuratList object}

\item{params}{a BaseMerge object}

\item{...}{Additional arguments}
}
\description{
Validate a SeuratList object before merging
}
