% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BaseMerge-class}
\alias{BaseMerge-class}
\title{BaseMerge}
\description{
The BaseMerge class
}
\section{Slots}{

\describe{
\item{\code{batch}}{character name of batch in dataset metadata}

\item{\code{dimreduc_names}}{name of dimension reduction in metadata, i.e. "PCA"}

\item{\code{return}}{data type to return, can be Seurat or SingleCellExperiment}

\item{\code{dependences}}{a list of dependend packages}

\item{\code{name}}{The name of the method used to store the dimension reduction results}

\item{\code{npcs}}{the size of latent dimension (default: 20L)}

\item{\code{latent}}{whether or not use latent representation to construct the neighboring graph (default: FALSE)}

\item{\code{umap_name}}{the name of the UMAP results}

\item{\code{umap_key}}{the name of the UMAP key}

\item{\code{umap_dim}}{the name of the UMAP dimension (default: 2L)}

\item{\code{snn_name}}{the name of SNN results}

\item{\code{knn_name}}{the name of the KNN results}
}}

