% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.r
\name{Cluster,Seurat,scLCACluster-method}
\alias{Cluster,Seurat,scLCACluster-method}
\title{Cluster a Seurat object by scLCA (https://bitbucket.org/scLCA/single_cell_lca/src/master/)}
\usage{
\S4method{Cluster}{Seurat,scLCACluster}(x, params, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{params}{a LeidenCluster object}

\item{...}{Additional arguments}
}
\value{
returns a data object with PCA embedding
}
\description{
Cluster a Seurat object by scLCA (https://bitbucket.org/scLCA/single_cell_lca/src/master/)
}
