% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblemerge.R
\name{EnsembleMerge}
\alias{EnsembleMerge}
\title{Run ensemblemerge functions}
\usage{
EnsembleMerge(
  data,
  methods = c("Seurat", "Harmony"),
  return = "Seurat",
  a = 1,
  b = 1,
  file = NA,
  normalization_method = "LogNormalize",
  latent = FALSE,
  ...
)
}
\arguments{
\item{x}{SummarizedExpirement object containing single cell counts matrix}
}
\value{
returns a SummarizedExperiment object of the integrated data
}
\description{
Run ensemblemerge functions
}
