% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblemerge.R
\name{ensemblemerge_core}
\alias{ensemblemerge_core}
\title{The core ensemble function}
\usage{
ensemblemerge_core(params, data)
}
\arguments{
\item{params}{a EnsembleMergeParams object}

\item{data}{a Seurat object}
}
\value{
a list of two elememts: (1) the intergrated cell-cell neighboring graph, and (2) weight of each constituent method
}
\description{
The core ensemble function
}
