% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HarmonyMerge-class}
\alias{HarmonyMerge-class}
\title{The HarmonyMerge class}
\description{
The HarmonyMerge class
}
\section{Slots}{

\describe{
\item{\code{theta_harmony}}{diversity clustering penalty parameter, larger values increase diversity}

\item{\code{num_clust}}{number of clusters to use in harmony integration}

\item{\code{max_iter_cluster}}{maximum number of learning iterations per cluster}
}}

