% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.r
\name{Merge,SeuratList,scVIMerge-method}
\alias{Merge,SeuratList,scVIMerge-method}
\title{Merge Seurat objects by using scVI}
\usage{
\S4method{Merge}{SeuratList,scVIMerge}(x, params, ...)
}
\arguments{
\item{x}{SeuratList object containing single cell counts matrix}

\item{params}{a scVIMerge object}

\item{...}{Additional arguments}
}
\value{
returns a Seurat object with integrated data
}
\description{
Merge Seurat objects by using scVI
}
\references{
https://scarches.readthedocs.io/en/latest/reference_building_from_scratch.html

Lopez, R., Regier, J., Cole, M.B. et al. Deep generative modeling for single-cell transcriptomics. Nat Methods 15, 1053–1058 (2018). https://doi.org/10.1038/s41592-018-0229-2
}
