% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.r
\name{Merge,SeuratList,SeuratMerge-method}
\alias{Merge,SeuratList,SeuratMerge-method}
\title{Batch correction by Seurat integration pipeline (https://satijalab.org/seurat/articles/integration_introduction.html)}
\usage{
\S4method{Merge}{SeuratList,SeuratMerge}(x, params, ...)
}
\arguments{
\item{x}{a SeuratList object}

\item{params}{a SeuratMerge object}

\item{...}{Additional arguments}
}
\value{
returns a data object with integrated data
}
\description{
Batch correction by Seurat integration pipeline (https://satijalab.org/seurat/articles/integration_introduction.html)
}
