% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{Preprocess}
\alias{Preprocess}
\title{The generic function of Preprocess}
\usage{
Preprocess(params, data, ...)
}
\arguments{
\item{params}{a parameter object}

\item{data}{a data object}

\item{...}{Other arguments}
}
\description{
The generic function of Preprocess
}
