% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBBKNN.R
\name{run_BBKNN}
\alias{run_BBKNN}
\title{Run BBKNN integration}
\usage{
run_BBKNN(params, data)
}
\arguments{
\item{params}{a scVIParams object}

\item{data}{a Seurat object}
}
\value{
returns a Seurat object with integrated data
}
\description{
Adopted from https://scanpy-tutorials.readthedocs.io/en/latest/integrating-data-using-ingest.html#BBKNN
}
