% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBBKNN.R
\name{run_BBKNN}
\alias{run_BBKNN}
\title{Run Seurat V3 functions}
\usage{
run_BBKNN(params, data)
}
\arguments{
\item{params}{BBKNNParams object generated from setParams(method = "BBKNN") function}

\item{data}{SingleCellExperiment object containing single cell counts matrix}
}
\value{
returns a SummarizedExperiment object of the integrated data
}
\description{
Run Seurat V3 functions
}
