% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMNN.R
\name{run_fastMNN}
\alias{run_fastMNN}
\title{Run fastMNN merging}
\usage{
run_fastMNN(params, data)
}
\arguments{
\item{params}{a FastMNNParams object}

\item{data}{a data object}
}
\value{
a Seurat object
}
\description{
Run fastMNN merging
}
