% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMNN.R
\name{run_fastMNN}
\alias{run_fastMNN}
\title{Run Seurat V3 scaling and normalizing}
\usage{
run_fastMNN(params, data)
}
\arguments{
\item{params}{FastMNNParams object generated from setParams(method="fastMNN", ...)}

\item{data}{SummarizedExpirement object containing single cell counts matrix}
}
\value{
returns a SummarizedExperiment object of the integrated data
}
\description{
Run Seurat V3 scaling and normalizing
}
