% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLiger.R
\name{run_Liger}
\alias{run_Liger}
\title{Run Seurat V3 functions}
\usage{
run_Liger(params, data)
}
\arguments{
\item{params}{LigerParams object}

\item{data}{SingleCellExpirement object containing single cell counts matrix}
}
\value{
returns a SingleCellExperiment object of the integrated data
}
\description{
Run Seurat V3 functions
}
