% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runScanorama.R
\name{run_Scanorama}
\alias{run_Scanorama}
\title{Run Seurat V3 functions}
\usage{
run_Scanorama(params, data)
}
\arguments{
\item{params}{ScanoramaParams object generated from setParams(method = "BBKNN") function}

\item{data}{SingleCellExperiment object containing single cell counts matrix}
}
\value{
returns a SummarizedExperiment object of the integrated data
}
\description{
Run Seurat V3 functions
}
