% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCVI.R
\name{run_scVI}
\alias{run_scVI}
\title{Run scVI functions}
\usage{
run_scVI(params, data)
}
\arguments{
\item{x}{SummarizedExpirement, SingleCellExperiment, or Seurat object containing single cell counts matrix}
}
\value{
returns a SummarizedExperiment object of the integrated data
}
\description{
Run scVI functions
}
