% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUncorrected.R
\name{run_Uncorrected}
\alias{run_Uncorrected}
\title{Run uncorrected merging}
\usage{
run_Uncorrected(params, data)
}
\arguments{
\item{params}{a UncorrectedMerge object}

\item{data}{a data object}
}
\value{
returns a Seurat object of the integrated data
}
\description{
Run uncorrected merging
}
