% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParams.R
\name{setParams}
\alias{setParams}
\title{set parameters for integration}
\usage{
setParams(methods = "Seurat", ...)
}
\arguments{
\item{method}{character specifying the integration method i.e "Seurat"}
}
\value{
returns a params S4 class depending on input
}
\description{
set parameters for integration
}
