% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SeuratMerge-class}
\alias{SeuratMerge-class}
\title{SeuratMerge}
\description{
class for merging dataset with Seurat
}
\section{Slots}{

\describe{
\item{\code{k.weight}}{weight for neighbor function}
}}

